+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Export the generated palettes from the palette dialog as a GIMP palette file
+
 ### Changed
 - Fixed a bug where the illuminants where not fully shown
 - The names of the GNOME color palette are now available